pyproject.toml
[tool.poetry]
name = "tuatara"
version = "0.1.0"
description = "Tuatara is a text-mode music player."
authors = ["Bill Nottingham <notting@splat.cc>"]
license = "GPL-3.0+"
readme = "README.md"
packages = [ { include = "caca" }, { include = "tuatara" } ]
include = ["sample-tuatara.toml"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pillow = "^10.1.0"
urllib3 = "^2.1.0"
pygobject = "^3.46.0"
m3u8 = "^4.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pex = "^2.1.154"
mock = "^5.1.0"

[tool.poetry.scripts]
tuatara = 'tuatara.main:main'

[tool.ruff]
exclude = ["caca"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
